TestServlet: changes   * replace numbers format ( int > float );   * add project file TestServlet.iml;   * change maven deploy configuration

+++ TestServlet/src/main/java/com.epolsoft.servlet/Common.java
@@ -1,11 +1,11 @@
 package com.epolsoft.servlet;
 
 public class Common {
-    public static int getNumber( String num_str ) throws NumberFormatException {
-        return Integer.parseInt( num_str );
+    public static float getNumber( String num_str ) throws NumberFormatException {
+        return Float.parseFloat( num_str );
     }
 
-    public static int Calculate( int n1, int n2 ) {
+    public static float Calculate( float n1, float n2 ) {
         return n1 + n2;
     }
 }
